Fix deserialization for second api

+++ server/StorMi/Interfaces/IGeoAPIHandler.cs
@@ -5,6 +5,6 @@ namespace StorMi.Interfaces
 {
     public interface IGeoAPIHandler
     {
-        public Task<IEnumerable<int>> GetCityCoordsByName(string cityName);
+        public Task<IEnumerable<double>> GetCityCoordsByName(string cityName);
     }
 }
